refactor, add test
